requirements.txt
Flask
flask-mail
mysql-connector-python
gunicorn
